# src/main/resources/static/openapi.yml
openapi: 3.0.1
info:
  title: YAPP20-BE-ToGaether API
  description: 모임 기반 플랫폼 ToGaether API 문서
  license:
    name: Apache License Version 2.0
    url: '[http://www.apache.org/licenses/LICENSE-2.0](http://www.apache.org/licenses/LICENSE-2.0)'
  version: '1.0'
servers:
  - url: https://yapp-togather.com
...
paths:
...
  /api/clubs/search:
    get:
      tags:
        - 모임 (Club)
      summary: 모임 검색 및 필터링
      description: Query String을 통해 데이터를 요청받고, 조건에 맞는 모임을 반환합니다.
      parameters:
        - name: request
          in: query
          description: SearchingRequest
          required: true
          schema:
            $ref: '#/components/schemas/SearchingRequest'
        - name: searchingType
          in: query
          description: 검색어를 통한 검색은 word, 카테고리 검색은 category
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchingResponse'
        '401':
          description: 토큰이 없거나 문제있는 경우에 아래와 같은 포맷으로 응답합니다. (아래는 토큰이 만료된 경우 예시입니다.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseInfo'
  /api/clubs/search/range:
    get:
      tags:
        - 모임 (Club)
      summary: 주어진 범위 내에 있는 모임 조회
      description: 좌상단, 우하단 위도 경도를 통해 해당 범위 내에 있는 모임을 조회합니다
      operationId: searchingWithinRange
      parameters:
        - name: request
          in: query
          description: SearchingWithinRangeClubRequest
          required: true
          schema:
            $ref: '#/components/schemas/SearchingWithinRangeClubRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchingWithinRangeClubResponse'
        '401':
          description: 토큰이 없거나 문제있는 경우에 아래와 같은 포맷으로 응답합니다. (아래는 토큰이 만료된 경우 예시입니다.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseInfo'
  /api/clubs/search/simple/{club-id}:
    get:
      tags:
        - 모임 (Club)
      summary: FIN 클릭 시 간단한 모임 조회
      description: 유저의 위도, 경도, 모임의 식별자 값을 통해 모임의 간단한 정보를 조회합니다
      parameters:
        - name: request
          in: query
          description: SearchingSimpleClubRequest
          required: true
          schema:
            $ref: '#/components/schemas/SearchingSimpleClubRequest'
        - name: club-id
          in: path
          description: 모임의 식별자 값
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchingSimpleClubResponse'
        '401':
          description: 토큰이 없거나 문제있는 경우에 아래와 같은 포맷으로 응답합니다. (아래는 토큰이 만료된 경우 예시입니다.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionResponseInfo'
components:
  schemas:
...
    ExceptionResponseInfo:
      type: object
      properties:
        status:
          type: string
          example: 401
        message:
          type: string
          example: 만료된 JWT토큰입니다.


    ############################################# Account #############################################
...
    SearchingRequest:
      type: object
      properties:
        searchingWord:
          title: 검색어
          type: string
          description: 카테고리를 통한 검색일 경우, 해당 값은 null
        category:
          type: string
          enum:
            - WALK
            - DOG_CAFE
        eligibleBreed:
          type: string
          example: "웰시코기"
        petSizeType:
          type: string
          enum:
            - LARGE
            - MEDIUM
            - SMALL
            - ALL
        eligibleSex:
          type: string
          enum:
            - MAN
            - WOMAN
            - ALL
        participateMin:
          title: 최소 참여 인원 필터링
          type: integer
          description: 3명 이하일 경우 해당 값은 0으로 전송
          format: int32
        participateMax:
          title: 최대 참여 인원 필터링
          type: integer
          format: int32
        page:
          type: integer
          description: 몇 번째 게시물인지 나타낸다
          format: int32
        startLatitude:
          type: number
          description: 사용자 위도
          format: double
        startLongitude:
          type: number
          description: 사용자 경도
          format: double
        status:
          type: string
          enum:
            - AVAILABLE
            - PERSONNEL_FULL
            - END
      description: 모임 조회 및 필터링 요청 API
    SearchingResponse:
      type: array
      items:
        properties:
          category:
            type: string
            enum:
              - WALK
              - DOG_CAFE
          title:
            type: string
            example: "산책할사람!"
          startDate:
            type: string
            format: date-time
          endDate:
            type: string
            format: date-time
          eligiblePetSizeTypes:
            type: array
            items:
              type: string
            example:
              - LARGE
              - MEDIUM
          eligibleBreeds:
            type: array
            items:
              type: string
            example:
              - 말티즈
              - 웰시코기
              - 리트리버
          eligibleSex:
            type: string
            example: "ALL"
          maximumPeople:
            type: integer
            example: "6"
          participants:
            type: integer
            example: "5"
          latitude:
            type: number
            format: double
            example: "2,174"
          longitude:
            type: number
            format: double
            example: "41,4033803"
          meetingPlace:
            type: string
            example: "서울 강남역 10번 출구"
          distance:
            type: integer
            example: "233m"
          status:
            type: string
            enum:
              - AVAILABLE
              - PERSONNEL_FULL
              - END
    SearchingWithinRangeClubRequest:
      type: object
      properties:
        upperLeftLatitude:
          title: 좌상단 위도
          type: number
          format: double
        upperLeftLongitude:
          title: 좌상단 경도
          type: number
          format: double
        bottomRightLatitude:
          title: 우하단 위도
          type: number
          format: double
        bottomRightLongitude:
          title: 우하단 경도
          type: number
          format: double
    SearchingWithinRangeClubResponse:
      type: array
      items:
        properties:
          clubId:
            type: integer
            description: 모임 식별자 값
            format: int64
          category:
            type: string
            enum:
              - WALK
              - DOG_CAFE
          clubLatitude:
            title: 모임 위도
            type: number
            format: double
          clubLongitude:
            title: 모임 경도
            type: number
            format: double
    SearchingSimpleClubRequest:
      type: object
      properties:
        userLatitude:
          title: 유저 위도
          type: number
          format: double
        userLongitude:
          title: 유저 경도
          type: number
          format: double
    SearchingSimpleClubResponse:
      type: object
      properties:
        category:
          type: string
          enum:
            - WALK
            - DOG_CAFE
        title:
          type: string
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        eligiblePetSizeTypes:
          type: array
          items:
            type: string
          example:
            - LARGE
            - MEDIUM
        eligibleBreeds:
          type: array
          items:
            type: string
          example:
            - 말티즈
            - 웰시코기
            - 리트리버
        eligibleSex:
          type: string
        maximumPeople:
          type: integer
        participants:
          type: integer
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        meetingPlace:
          type: string
        distance:
          type: number
          format: double
        clubStatus:
          type: string
          enum:
            - AVAILABLE
            - END